Add msgpack serializer

+++ Sources/Wampproto/Messages/Message.swift
@@ -1,7 +1,7 @@
 import Foundation
 
 protocol Message {
-    var type: Int { get }
+    var type: Int64 { get }
     static func parse(message: [Any])throws -> Message
     func marshal() -> [Any]
 }
